Making the game
Actually got game working, can now generate the random number that needs to be guessed and can read the inputs of the user to determine if the answer written in the forms is right or wrong. Changes background color accordingly to show right or wrong. Next iteration will include higher or lower colors to make guessing easier